// types.ts
export interface Question {
  id: number;
  text: string;
  options: string[];
  correctAnswerIndex: number;
  prize: string;
  difficulty: 'easy' | 'medium' | 'hard';
  subject: string;
  chapter: string;
  lesson: string;
}
// ...
export interface Lifelines {
  fiftyFifty: boolean;
  askAudience: boolean;
  callFriend: boolean;
}

export interface MoneyTier {
  level: number;
  amount: string;
  isSafeHaven: boolean;
  value: number;
}